Finished the course
Renamed some files in an effort to make them more intuitive. Finished the course which includes; uplaoding the files to google photos, using the google photo links in place of the local paths for the images so they work in an actual email. Adding alt text to images so the email looks alright when images are disabled. The email is also now an email template in gmail for me (but that is not shown in this repo)